+++ assets/ap_start.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 rfkill unblock wifi
 
-ip link set Lan up
+ip link set CHANGE_ME_LAN up
 
 docker rm -f hotspot 2>/dev/null
 
@@ -10,9 +10,9 @@ docker run --rm \
     --cap-add=NET_ADMIN \
     --cap-add=NET_RAW \
     --privileged \
-    -e WLAN_INT="Lan" \
+    -e WLAN_INT="CHANGE_ME_LAN" \
     -e WLAN_IP="192.168.200.1" \
     -e WLAN_MASK="24" \
     -v /etc/ap_config/:/etc/conf \
     --name hotspot \
-    mrdgidgi/simple-hotspot
+    mrdgidgi/simple-hotspot:VERSION_simpleHotspot